端口模块设置修改：使用调节阀控制水泵。 控制协议增加水泵调节。 在检定台实验通过：阀门控制、水泵控制、温度采集、天平采集
git-svn-id: https://192.168.1.124:8443/svn/delusoft/trunk/qtverify@172 62fcab80-fd93-7748-8c49-e77283212eb3

+++ source/include/protocol.h
@@ -149,7 +149,7 @@ public:
 
 public slots:
 	void makeRelaySendBuf(UINT8 portno, bool status);
-	void makeRegulateSendBuf(UINT8 portno, int degree);
+	void makeRegulateSendBuf(UINT8 portno, UINT16 degree);
 	void makeQuerySendBuf();
 	QByteArray getSendBuf();
 	UINT8 readControlComBuffer(QByteArray tmp);
